@@ -21,17 +21,17 @@
     }
   ],
   "require": {
-    "php": ">=8.0",
-    "illuminate/database": "~9",
-    "illuminate/support": "~9",
+    "php": ">=8.1",
+    "illuminate/database": "~10",
+    "illuminate/support": "~10",
     "laravel/legacy-factories": "^1.3.2"
   },
   "require-dev": {
     "mockery/mockery": "^1.6",
-    "orchestra/testbench": "^7.0",
+    "orchestra/testbench": "^8.0",
     "phpmd/phpmd": "^2.6|^2.7|^2.8|^2.9|^2.15",
     "phpunit/phpunit": "^7.3|^8.0|^9.0|^10",
-    "illuminate/database": "~9"
+    "illuminate/database": "~10"
   },
   "autoload": {
     "classmap": [
